Refactoring: simplify build menu functions
Note that sub menu items now follow the same rules as the main items for
activation. This should not have any impact on the original C3 behavior
but this introduce a slight change in the way Julius handles hotkeys in
a very rare situation: during the first 2 tutorial levels, you cannot
clone a sub-item building anymore (like a small temple) unless they can
be built from the build menu.