Revert "Hide resources that cannot be traded in empire screen"
This commit also hid the resources that cities have stopped trading,
even though they still appear in the trade advisor and we can still
technically export/import it. If we did not open the trade route, this
means that a new resource could be added to the trade advisor without
knowing where it comes from.